MIDI playback (#44)
This PR replaces OSMD's audio player with MIDI playback. The MIDI file is generated on the fly from the MusicXML via `musicxml-mma` API endpoint. Playback is done via `midi-player` and associated modules.

To keep the sheet in sync with the playback, the MIDI file includes measure markers. The OSMD cursor is manually positioned to highlight the current notes based on current measure + relative timestamp inside the measure. This saves us from having to recompute repeats and jumps on the score side, since the MIDI file has already unrolled the repeats and jumps into a linear sequence of notes.

This approach also allows maximum flexibility in reusing MMA to generate an accompaniment that is agnostic to the music engraver. The only interface between score rendering and playback is the ability to sync the position from one to the other.